Fix: buffer ValueSerializer creates from huge object undeserializable
1. ValueSerializer buffer limitation now changed from 2GB to 4GB
2. Buffers larger than 2GB are now deserializable
3. Update AUTHORS

Ref: https://github.com/nodejs/node/issues/40059

+++ src/api/api.cc
@@ -3380,10 +3380,10 @@ ValueDeserializer::ValueDeserializer(Isolate* isolate, const uint8_t* data,
 
 ValueDeserializer::ValueDeserializer(Isolate* isolate, const uint8_t* data,
                                      size_t size, Delegate* delegate) {
-  if (base::IsValueInRangeForNumericType<int>(size)) {
+  if (base::IsValueInRangeForNumericType<size_t>(size)) {
     private_ = new PrivateData(
         reinterpret_cast<i::Isolate*>(isolate),
-        base::Vector<const uint8_t>(data, static_cast<int>(size)), delegate);
+        base::Vector<const uint8_t>(data, static_cast<size_t>(size)), delegate);
   } else {
     private_ =
         new PrivateData(reinterpret_cast<i::Isolate*>(isolate),
